Holberton Task

+++ structures_typedef/dog.h
@@ -1,6 +1,6 @@
 #ifndef MAIN_H
 #define MAIN_H
 
-struct dog{char* name; char* owner; int age;};
+struct dog{char* name; char* owner; float age;};
 
 #endif
